Fix circular dependency in voice streamed example by renaming agents.py to my_workflow.py
